Fallback filename when Content-Disposition header is missing.
The `Content-Disposition` header is expected to be present when
downloading a sheet as a file. That expectation appears to no longer be
always true.

The question of whether or not we should be able to expect the
`Content-Disposition` header has been raised internally. In the
meantime, the user will get a default filename of `download`, along with
the appropriate suffix when we receive an expected `Content-Type`
header.

Ref: https://github.com/smartsheet/smartsheet-python-sdk/issues/67